Fix /addpoint requiring relog + widen Life/Mana/Shield/BP overflow above 65000
Ported from project_mu_main52_63, matching its already-verified fix.

Server (Protocol.h/.cpp, DSProtocol.cpp, MasterSkillTree.h/.cpp, CommandManager.cpp):
- /v /a /s /e /c commands now call the new GCLevelUpPointSend (the same
  C1:F3:06 confirmation the native "+" button uses) instead of
  GCNewCharacterInfoSend, which is a no-op when GAMESERVER_EXTRA==0 (this
  build) - that gap is why added points only ever showed up after
  relogging. Sent as ONE packet carrying the full amount instead of one
  packet per point, so large amounts (e.g. /a 1000) no longer flood/freeze
  the client.
- PMSG_LIFE_SEND/PMSG_MANA_SEND (the packet sent on every damage/heal/regen
  tick) and PMSG_CHARACTER_INFO_SEND, PMSG_CHARACTER_REGEN_SEND,
  PMSG_LEVEL_UP_SEND, PMSG_MASTER_INFO_SEND, PMSG_MASTER_LEVEL_UP_SEND all
  had their Life/Mana/Shield/BP/LevelUpPoint (and Max* variants) widened
  WORD->DWORD, GET_MAX_WORD_VALUE clamps removed at every fill site - these
  no longer wrap/cap at 65000/65535.

Client (_struct.h, WSclient.h/.cpp, CGFxMainUi.cpp, NewUIMainFrameWindow.cpp,
SocketSystem.cpp):
- CHARACTER_ATTRIBUTE/MASTER_LEVEL_VALUE fields and matching receive structs
  widened to match.
- ReceiveAddPoint applies += amount (was ++), reads the new wide
  MaxLifeAndManaWide field instead of the clamped Max.
- ReceiveLife/ReceiveMana reassemble 4-byte values instead of 2-byte
  (PRECEIVE_LIFE.Life[5]->Life[9] to fit both packet layouts).
- 3 UI files had local WORD copies in the HP/MP/SD/AG gauge-calculation
  functions bumped to DWORD - they were truncating the already-correct wide
  value right back down.

Left out of scope (unrelated bugs that showed up in the diff against the
reference project, from other sessions there): mailbox letter crash fix,
guild notice overflow, master skill tree list caching/sync, teleport-hack
detection, IP logging. Can port those separately if wanted.

Confirmed working in-game by Evan.

+++ Source_Client/source/_struct.h
@@ -385,12 +385,14 @@ typedef struct
 	WORD Vitality;
 	WORD Energy;
 	WORD Charisma;
-	WORD Life;
-	WORD Mana;
-	WORD LifeMax;
-	WORD ManaMax;
-	WORD Shield;
-	WORD ShieldMax;
+	// Widened WORD->DWORD, same overflow reason as LevelUpPoint above: with
+	// Vitality/Energy maxed, these blow way past 65535.
+	DWORD Life;
+	DWORD Mana;
+	DWORD LifeMax;
+	DWORD ManaMax;
+	DWORD Shield;
+	DWORD ShieldMax;
 	WORD AttackRatingPK;
 	WORD SuccessfulBlockingPK;
 	WORD AddStrength;
@@ -400,8 +402,10 @@ typedef struct
 	WORD AddLifeMax;
 	WORD AddManaMax;
 	WORD AddCharisma;
-	WORD SkillMana;
-	WORD SkillManaMax;
+	// SkillMana/SkillManaMax are this client's naming for server's BP/MaxBP.
+	// Widened WORD->DWORD, same overflow reason as Life/Mana above.
+	DWORD SkillMana;
+	DWORD SkillManaMax;
 	BYTE Ability;
 	WORD AbilityTime[3];
 
@@ -425,7 +429,11 @@ typedef struct
 	WORD Defense;
 	WORD MagicDefense;
 	WORD WalkSpeed;
-	WORD LevelUpPoint;
+	// Widened WORD->DWORD -- a character with all 5 stats maxed at 32767 plus
+	// many resets can accrue well over 65535 unspent points to distribute,
+	// which wrapped/rendered wrong when this was a WORD (matches the widened
+	// LevelUpPoint fields in WSclient.h's PRECEIVE_JOIN_MAP_SERVER/PRECEIVE_LEVEL_UP).
+	DWORD LevelUpPoint;
 	BYTE SkillNumber;
 	BYTE SkillMasterNumber;
 	WORD Skill[MAX_SKILLS];
@@ -444,10 +452,10 @@ typedef struct _MASTER_LEVEL_VALUE
 	short		nTotalMPoint;
 	short		nMaxPoint;
 
-	WORD		wMaxLife;
-	WORD		wMaxMana;
-	WORD		wMaxShield;
-	WORD		wMaxBP;
+	DWORD		wMaxLife; // widened WORD->DWORD, same overflow reason as CHARACTER_ATTRIBUTE::LifeMax
+	DWORD		wMaxMana; // widened WORD->DWORD, same reason
+	DWORD		wMaxShield; // widened WORD->DWORD, same reason
+	DWORD		wMaxBP; // widened WORD->DWORD, same reason
 
 } MASTER_LEVEL_VALUE;
